nuc64: auto generate

+++ speedtest1/ccgo_linux_amd64.go
@@ -1708,7 +1708,7 @@ func HashInit(tls *libc.TLS) {
 		if !(k < uint32(256)) {
 			break
 		}
-		*(*uint8)(unsafe.Pointer(uintptr(unsafe.Pointer(&g)) + 3152 + 3 + uintptr(k))) = uint8(uint8(k))
+		*(*uint8)(unsafe.Pointer(uintptr(unsafe.Pointer(&g)) + 3152 + 3 + uintptr(k))) = uint8(k)
 		goto _1
 	_1:
 		;
@@ -1729,7 +1729,7 @@ func HashUpdate(tls *libc.TLS, aData uintptr, nData uint32) {
 	i = g.hash.i
 	j = g.hash.j
 	if g.hashFile != 0 {
-		libc.Xfwrite(tls, aData, uint64(1), uint64(uint64(nData)), g.hashFile)
+		libc.Xfwrite(tls, aData, uint64(1), uint64(nData), g.hashFile)
 	}
 	k = uint32(0)
 	for {
@@ -1769,7 +1769,7 @@ func HashFinal(tls *libc.TLS) {
 		}
 		i++
 		t = *(*uint8)(unsafe.Pointer(uintptr(unsafe.Pointer(&g)) + 3152 + 3 + uintptr(i)))
-		j = uint8(int32(j) + int32(int32(t)))
+		j = uint8(int32(j) + int32(t))
 		*(*uint8)(unsafe.Pointer(uintptr(unsafe.Pointer(&g)) + 3152 + 3 + uintptr(i))) = *(*uint8)(unsafe.Pointer(uintptr(unsafe.Pointer(&g)) + 3152 + 3 + uintptr(j)))
 		*(*uint8)(unsafe.Pointer(uintptr(unsafe.Pointer(&g)) + 3152 + 3 + uintptr(j))) = t
 		t = uint8(int32(t) + int32(*(*uint8)(unsafe.Pointer(uintptr(unsafe.Pointer(&g)) + 3152 + 3 + uintptr(i)))))
@@ -1791,14 +1791,14 @@ func HashFinal(tls *libc.TLS) {
 //	** is not a hex digit.
 //	*/
 func hexDigitValue(tls *libc.TLS, c int8) (r int32) {
-	if int32(int32(c)) >= int32('0') && int32(int32(c)) <= int32('9') {
-		return int32(int32(c)) - int32('0')
+	if int32(c) >= int32('0') && int32(c) <= int32('9') {
+		return int32(c) - int32('0')
 	}
-	if int32(int32(c)) >= int32('a') && int32(int32(c)) <= int32('f') {
-		return int32(int32(c)) - int32('a') + int32(10)
+	if int32(c) >= int32('a') && int32(c) <= int32('f') {
+		return int32(c) - int32('a') + int32(10)
 	}
-	if int32(int32(c)) >= int32('A') && int32(int32(c)) <= int32('F') {
-		return int32(int32(c)) - int32('A') + int32(10)
+	if int32(c) >= int32('A') && int32(c) <= int32('F') {
+		return int32(c) - int32('A') + int32(10)
 	}
 	return -int32(1)
 }
@@ -1834,7 +1834,7 @@ func integerValue(tls *libc.TLS, zArg uintptr) (r int32) {
 			if !(v1 >= 0) {
 				break
 			}
-			v = v<<libc.Int32FromInt32(4) + int64(int64(x))
+			v = v<<libc.Int32FromInt32(4) + int64(x)
 			zArg++
 		}
 	} else {
@@ -1845,7 +1845,7 @@ func integerValue(tls *libc.TLS, zArg uintptr) (r int32) {
 	}
 	i = 0
 	for {
-		if !(uint64(uint64(i)) < libc.Uint64FromInt64(144)/libc.Uint64FromInt64(16)) {
+		if !(uint64(i) < libc.Uint64FromInt64(144)/libc.Uint64FromInt64(16)) {
 			break
 		}
 		if libsqlite3.Xsqlite3_stricmp(tls, aMult[i].zSuffix, zArg) == 0 {
@@ -2398,7 +2398,7 @@ func speedtest1_run(tls *libc.TLS) {
 						}
 						g.nResByte += uint64(nBlob*int32(2) + int32(2))
 					} else {
-						HashUpdate(tls, z1, uint32(uint32(len1)))
+						HashUpdate(tls, z1, uint32(len1))
 						g.nResByte += uint64(len1 + int32(2))
 					}
 				}
@@ -2509,7 +2509,7 @@ func testset_main(tls *libc.TLS) {
 	n = v1
 	sz = v1
 	(*(*[2000]int8)(unsafe.Pointer(bp)))[0] = 0
-	maxb = int32(roundup_allones(tls, uint32(uint32(sz))))
+	maxb = int32(roundup_allones(tls, uint32(sz)))
 	speedtest1_begin_test(tls, int32(100), __ccgo_ts+590, libc.VaList(bp+2008, n))
 	speedtest1_exec(tls, __ccgo_ts+626, 0)
 	speedtest1_exec(tls, __ccgo_ts+632, libc.VaList(bp+2008, isTemp(tls, int32(9)), g.zNN, g.zNN, g.zNN))
@@ -2519,9 +2519,9 @@ func testset_main(tls *libc.TLS) {
 		if !(i <= n) {
 			break
 		}
-		x1 = swizzle(tls, uint32(uint32(i)), uint32(uint32(maxb)))
+		x1 = swizzle(tls, uint32(i), uint32(maxb))
 		speedtest1_numbername(tls, x1, bp, int32(2000))
-		libsqlite3.Xsqlite3_bind_int64(tls, g.pStmt, int32(1), int64(int64(x1)))
+		libsqlite3.Xsqlite3_bind_int64(tls, g.pStmt, int32(1), int64(x1))
 		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(2), i)
 		libsqlite3.Xsqlite3_bind_text(tls, g.pStmt, int32(3), bp, -int32(1), libc.UintptrFromInt32(0))
 		speedtest1_run(tls)
@@ -2542,10 +2542,10 @@ func testset_main(tls *libc.TLS) {
 		if !(i <= n) {
 			break
 		}
-		x1 = swizzle(tls, uint32(uint32(i)), uint32(uint32(maxb)))
+		x1 = swizzle(tls, uint32(i), uint32(maxb))
 		speedtest1_numbername(tls, x1, bp, int32(2000))
 		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), i)
-		libsqlite3.Xsqlite3_bind_int64(tls, g.pStmt, int32(2), int64(int64(x1)))
+		libsqlite3.Xsqlite3_bind_int64(tls, g.pStmt, int32(2), int64(x1))
 		libsqlite3.Xsqlite3_bind_text(tls, g.pStmt, int32(3), bp, -int32(1), libc.UintptrFromInt32(0))
 		speedtest1_run(tls)
 		goto _3
@@ -2565,10 +2565,10 @@ func testset_main(tls *libc.TLS) {
 		if !(i <= n) {
 			break
 		}
-		x1 = swizzle(tls, uint32(uint32(i)), uint32(uint32(maxb)))
+		x1 = swizzle(tls, uint32(i), uint32(maxb))
 		speedtest1_numbername(tls, x1, bp, int32(2000))
 		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(2), i)
-		libsqlite3.Xsqlite3_bind_int64(tls, g.pStmt, int32(1), int64(int64(x1)))
+		libsqlite3.Xsqlite3_bind_int64(tls, g.pStmt, int32(1), int64(x1))
 		libsqlite3.Xsqlite3_bind_text(tls, g.pStmt, int32(3), bp, -int32(1), libc.UintptrFromInt32(0))
 		speedtest1_run(tls)
 		goto _4
@@ -2588,11 +2588,11 @@ func testset_main(tls *libc.TLS) {
 			break
 		}
 		if (i-int32(1))%g.nRepeat == 0 {
-			x1 = speedtest1_random(tls) % uint32(uint32(maxb))
+			x1 = speedtest1_random(tls) % uint32(maxb)
 			x2 = speedtest1_random(tls)%uint32(10) + uint32(sz/int32(5000)) + x1
 		}
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(int32(x1)))
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(2), int32(int32(x2)))
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(x1))
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(2), int32(x2))
 		speedtest1_run(tls)
 		goto _5
 	_5:
@@ -2611,9 +2611,9 @@ func testset_main(tls *libc.TLS) {
 			break
 		}
 		if (i-int32(1))%g.nRepeat == 0 {
-			x1 = speedtest1_random(tls) % uint32(uint32(maxb))
+			x1 = speedtest1_random(tls) % uint32(maxb)
 			(*(*[2000]int8)(unsafe.Pointer(bp)))[0] = int8('%')
-			len1 = speedtest1_numbername(tls, uint32(uint32(i)), bp+uintptr(1), int32(libc.Uint64FromInt64(2000)-libc.Uint64FromInt32(2)))
+			len1 = speedtest1_numbername(tls, uint32(i), bp+uintptr(1), int32(libc.Uint64FromInt64(2000)-libc.Uint64FromInt32(2)))
 			(*(*[2000]int8)(unsafe.Pointer(bp)))[len1] = int8('%')
 			(*(*[2000]int8)(unsafe.Pointer(bp)))[len1+int32(1)] = 0
 		}
@@ -2636,9 +2636,9 @@ func testset_main(tls *libc.TLS) {
 			break
 		}
 		if (i-int32(1))%g.nRepeat == 0 {
-			x1 = speedtest1_random(tls) % uint32(uint32(maxb))
+			x1 = speedtest1_random(tls) % uint32(maxb)
 			(*(*[2000]int8)(unsafe.Pointer(bp)))[0] = int8('%')
-			len1 = speedtest1_numbername(tls, uint32(uint32(i)), bp+uintptr(1), int32(libc.Uint64FromInt64(2000)-libc.Uint64FromInt32(2)))
+			len1 = speedtest1_numbername(tls, uint32(i), bp+uintptr(1), int32(libc.Uint64FromInt64(2000)-libc.Uint64FromInt32(2)))
 			(*(*[2000]int8)(unsafe.Pointer(bp)))[len1] = int8('%')
 			(*(*[2000]int8)(unsafe.Pointer(bp)))[len1+int32(1)] = 0
 		}
@@ -2661,9 +2661,9 @@ func testset_main(tls *libc.TLS) {
 			break
 		}
 		if (i-int32(1))%g.nRepeat == 0 {
-			x1 = speedtest1_random(tls) % uint32(uint32(maxb))
+			x1 = speedtest1_random(tls) % uint32(maxb)
 			(*(*[2000]int8)(unsafe.Pointer(bp)))[0] = int8('%')
-			len1 = speedtest1_numbername(tls, uint32(uint32(i)), bp+uintptr(1), int32(libc.Uint64FromInt64(2000)-libc.Uint64FromInt32(2)))
+			len1 = speedtest1_numbername(tls, uint32(i), bp+uintptr(1), int32(libc.Uint64FromInt64(2000)-libc.Uint64FromInt32(2)))
 			(*(*[2000]int8)(unsafe.Pointer(bp)))[len1] = int8('%')
 			(*(*[2000]int8)(unsafe.Pointer(bp)))[len1+int32(1)] = 0
 		}
@@ -2695,11 +2695,11 @@ func testset_main(tls *libc.TLS) {
 			break
 		}
 		if (i-int32(1))%g.nRepeat == 0 {
-			x1 = speedtest1_random(tls) % uint32(uint32(maxb))
+			x1 = speedtest1_random(tls) % uint32(maxb)
 			x2 = speedtest1_random(tls)%uint32(10) + uint32(sz/int32(5000)) + x1
 		}
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(int32(x1)))
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(2), int32(int32(x2)))
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(x1))
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(2), int32(x2))
 		speedtest1_run(tls)
 		goto _9
 	_9:
@@ -2718,11 +2718,11 @@ func testset_main(tls *libc.TLS) {
 			break
 		}
 		if (i-int32(1))%g.nRepeat == 0 {
-			x1 = speedtest1_random(tls) % uint32(uint32(maxb))
+			x1 = speedtest1_random(tls) % uint32(maxb)
 			x2 = speedtest1_random(tls)%uint32(10) + uint32(sz/int32(5000)) + x1
 		}
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(int32(x1)))
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(2), int32(int32(x2)))
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(x1))
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(2), int32(x2))
 		speedtest1_run(tls)
 		goto _10
 	_10:
@@ -2741,7 +2741,7 @@ func testset_main(tls *libc.TLS) {
 			break
 		}
 		if (i-int32(1))%g.nRepeat == 0 {
-			x1 = swizzle(tls, uint32(uint32(i)), uint32(uint32(maxb)))
+			x1 = swizzle(tls, uint32(i), uint32(maxb))
 			len1 = speedtest1_numbername(tls, x1, bp, int32(libc.Uint64FromInt64(2000)-libc.Uint64FromInt32(1)))
 		}
 		libsqlite3.Xsqlite3_bind_text(tls, g.pStmt, int32(1), bp, len1, libc.UintptrFromInt32(0))
@@ -2783,10 +2783,10 @@ func testset_main(tls *libc.TLS) {
 		if !(i <= n) {
 			break
 		}
-		x1 = speedtest1_random(tls) % uint32(uint32(maxb))
+		x1 = speedtest1_random(tls) % uint32(maxb)
 		x2 = speedtest1_random(tls)%uint32(10) + uint32(sz/int32(5000)) + x1
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(int32(x1)))
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(2), int32(int32(x2)))
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(x1))
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(2), int32(x2))
 		speedtest1_run(tls)
 		goto _12
 	_12:
@@ -2804,8 +2804,8 @@ func testset_main(tls *libc.TLS) {
 		if !(i <= n) {
 			break
 		}
-		x1 = speedtest1_random(tls)%uint32(uint32(sz)) + uint32(1)
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(int32(x1)))
+		x1 = speedtest1_random(tls)%uint32(sz) + uint32(1)
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(x1))
 		speedtest1_run(tls)
 		goto _13
 	_13:
@@ -2829,10 +2829,10 @@ func testset_main(tls *libc.TLS) {
 		if !(i <= n) {
 			break
 		}
-		x1 = speedtest1_random(tls)%uint32(uint32(maxb)) + uint32(1)
+		x1 = speedtest1_random(tls)%uint32(maxb) + uint32(1)
 		x2 = speedtest1_random(tls)%uint32(10) + uint32(sz/int32(5000)) + x1
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(int32(x1)))
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(2), int32(int32(x2)))
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(x1))
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(2), int32(x2))
 		speedtest1_run(tls)
 		goto _14
 	_14:
@@ -2850,8 +2850,8 @@ func testset_main(tls *libc.TLS) {
 		if !(i <= n) {
 			break
 		}
-		x1 = speedtest1_random(tls)%uint32(uint32(sz)) + uint32(1)
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(int32(x1)))
+		x1 = speedtest1_random(tls)%uint32(sz) + uint32(1)
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(x1))
 		speedtest1_run(tls)
 		goto _15
 	_15:
@@ -2878,10 +2878,10 @@ func testset_main(tls *libc.TLS) {
 		if !(i <= n) {
 			break
 		}
-		x1 = speedtest1_random(tls)%uint32(uint32(sz)) + uint32(1)
+		x1 = speedtest1_random(tls)%uint32(sz) + uint32(1)
 		x2 = speedtest1_random(tls)%uint32(10) + x1 + uint32(4)
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(int32(x1)))
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(2), int32(int32(x2)))
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(x1))
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(2), int32(x2))
 		speedtest1_run(tls)
 		goto _16
 	_16:
@@ -2909,9 +2909,9 @@ func testset_main(tls *libc.TLS) {
 		if !(i <= n) {
 			break
 		}
-		x1 = swizzle(tls, uint32(uint32(i)), uint32(uint32(maxb)))
-		speedtest1_numbername(tls, uint32(uint32(i)), bp, int32(2000))
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(int64(int64(x1))))
+		x1 = swizzle(tls, uint32(i), uint32(maxb))
+		speedtest1_numbername(tls, uint32(i), bp, int32(2000))
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(int64(x1)))
 		libsqlite3.Xsqlite3_bind_text(tls, g.pStmt, int32(2), bp, -int32(1), libc.UintptrFromInt32(0))
 		speedtest1_run(tls)
 		goto _18
@@ -2937,8 +2937,8 @@ func testset_main(tls *libc.TLS) {
 		if !(i <= n) {
 			break
 		}
-		x1 = swizzle(tls, uint32(uint32(i)), uint32(uint32(maxb)))
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(int64(int64(x1))))
+		x1 = swizzle(tls, uint32(i), uint32(maxb))
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(int64(x1)))
 		speedtest1_run(tls)
 		goto _19
 	_19:
@@ -2968,7 +2968,7 @@ func testset_main(tls *libc.TLS) {
 		if !(i <= n) {
 			break
 		}
-		x1 = swizzle(tls, uint32(uint32(i)), uint32(uint32(maxb)))
+		x1 = swizzle(tls, uint32(i), uint32(maxb))
 		speedtest1_numbername(tls, x1, bp, int32(2000))
 		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(2), i)
 		libsqlite3.Xsqlite3_bind_text(tls, g.pStmt, int32(1), bp, -int32(1), libc.UintptrFromInt32(0))
@@ -2991,7 +2991,7 @@ func testset_main(tls *libc.TLS) {
 		if !(i <= n) {
 			break
 		}
-		x1 = swizzle(tls, uint32(uint32(i)), uint32(uint32(maxb)))
+		x1 = swizzle(tls, uint32(i), uint32(maxb))
 		speedtest1_numbername(tls, x1, bp, int32(2000))
 		libsqlite3.Xsqlite3_bind_text(tls, g.pStmt, int32(1), bp, -int32(1), libc.UintptrFromInt32(0))
 		speedtest1_run(tls)
@@ -3212,13 +3212,13 @@ func testset_orm(tls *libc.TLS) {
 			case int32('I'):
 				fallthrough
 			case int32('T'):
-				libsqlite3.Xsqlite3_bind_int64(tls, g.pStmt, int32(j+uint32(2)), int64(int64(x1)))
+				libsqlite3.Xsqlite3_bind_int64(tls, g.pStmt, int32(j+uint32(2)), int64(x1))
 			case int32('F'):
-				libsqlite3.Xsqlite3_bind_double(tls, g.pStmt, int32(j+uint32(2)), float64(float64(x1)))
+				libsqlite3.Xsqlite3_bind_double(tls, g.pStmt, int32(j+uint32(2)), float64(x1))
 			case int32('V'):
 				fallthrough
 			case int32('B'):
-				libsqlite3.Xsqlite3_bind_text64(tls, g.pStmt, int32(j+uint32(2)), bp, uint64(uint64(len1)), libc.UintptrFromInt32(0), uint8(SQLITE_UTF8))
+				libsqlite3.Xsqlite3_bind_text64(tls, g.pStmt, int32(j+uint32(2)), bp, uint64(len1), libc.UintptrFromInt32(0), uint8(SQLITE_UTF8))
 				break
 			}
 			goto _3
@@ -3243,7 +3243,7 @@ func testset_orm(tls *libc.TLS) {
 			break
 		}
 		x1 = speedtest1_random(tls) % nRow
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(int32(x1)))
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(x1))
 		speedtest1_run(tls)
 		goto _4
 	_4:
@@ -3279,8 +3279,8 @@ func testset_trigger(tls *libc.TLS) {
 			if !(ii < NROW) {
 				break
 			}
-			x1 = int32(speedtest1_random(tls) % uint32(uint32(NROW)))
-			speedtest1_numbername(tls, uint32(uint32(x1)), bp, int32(2000))
+			x1 = int32(speedtest1_random(tls) % uint32(NROW))
+			speedtest1_numbername(tls, uint32(x1), bp, int32(2000))
 			libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), x1)
 			libsqlite3.Xsqlite3_bind_text(tls, g.pStmt, int32(2), bp, -int32(1), libc.UintptrFromInt32(0))
 			speedtest1_run(tls)
@@ -3422,7 +3422,7 @@ func testset_trigger(tls *libc.TLS) {
 		if !(jj < NROW2) {
 			break
 		}
-		speedtest1_numbername(tls, uint32(uint32(jj)), bp, int32(2000))
+		speedtest1_numbername(tls, uint32(jj), bp, int32(2000))
 		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), jj)
 		libsqlite3.Xsqlite3_bind_text(tls, g.pStmt, int32(2), bp, -int32(1), libc.UintptrFromInt32(0))
 		speedtest1_run(tls)
@@ -3488,7 +3488,7 @@ func testset_trigger(tls *libc.TLS) {
 		if !(jj < NROW2) {
 			break
 		}
-		speedtest1_numbername(tls, uint32(uint32(jj)), bp, int32(2000))
+		speedtest1_numbername(tls, uint32(jj), bp, int32(2000))
 		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), jj)
 		libsqlite3.Xsqlite3_bind_text(tls, g.pStmt, int32(2), bp, -int32(1), libc.UintptrFromInt32(0))
 		speedtest1_run(tls)
@@ -3742,11 +3742,11 @@ func displayLinuxIoStats(tls *libc.TLS, out uintptr) {
 	for libc.Xfgets(tls, bp, int32(200), in) != uintptr(0) {
 		i = 0
 		for {
-			if !(uint64(uint64(i)) < libc.Uint64FromInt64(112)/libc.Uint64FromInt64(16)) {
+			if !(uint64(i) < libc.Uint64FromInt64(112)/libc.Uint64FromInt64(16)) {
 				break
 			}
 			n = int32(libc.Xstrlen(tls, aTrans[i].zPattern))
-			if libc.Xstrncmp(tls, aTrans[i].zPattern, bp, uint64(uint64(n))) == 0 {
+			if libc.Xstrncmp(tls, aTrans[i].zPattern, bp, uint64(n)) == 0 {
 				libc.Xfprintf(tls, out, __ccgo_ts+19321, libc.VaList(bp+208, aTrans[i].zDesc, bp+uintptr(n)))
 				break
 			}
@@ -3834,9 +3834,9 @@ func main1(tls *libc.TLS, argc int32, argv uintptr) (r int32) {
 	doTrace = 0                                                                                      /* True for --trace */
 	zEncoding = uintptr(0)                                                                           /* --utf16be or --utf16le */
 	zDbName = uintptr(0)                                                                             /* Name of the test database */
-	pHeap = uintptr(0)   /* Allocated heap space */
-	pLook = uintptr(0)   /* Allocated lookaside space */
-	pPCache = uintptr(0) /* API return code */
+	pHeap = uintptr(0)                                                                               /* Allocated heap space */
+	pLook = uintptr(0)                                                                               /* Allocated lookaside space */
+	pPCache = uintptr(0)                                                                             /* API return code */
 	/*
 	 ** Confirms that argc has at least N arguments following argv[i]. */
 	/* Display the version of SQLite being tested */
@@ -4175,7 +4175,7 @@ func main1(tls *libc.TLS, argc int32, argv uintptr) (r int32) {
 		i++
 	}
 	if nHeap > 0 {
-		pHeap = libc.Xmalloc(tls, uint64(uint64(nHeap)))
+		pHeap = libc.Xmalloc(tls, uint64(nHeap))
 		if pHeap == uintptr(0) {
 			fatal_error(tls, __ccgo_ts+20093, libc.VaList(bp+16, nHeap))
 		}
@@ -4186,9 +4186,9 @@ func main1(tls *libc.TLS, argc int32, argv uintptr) (r int32) {
 	}
 	if doPCache != 0 {
 		if nPCache > 0 && szPCache > 0 {
-			pPCache = libc.Xmalloc(tls, uint64(int64(int64(nPCache))*int64(int64(szPCache))))
+			pPCache = libc.Xmalloc(tls, uint64(int64(nPCache)*int64(szPCache)))
 			if pPCache == uintptr(0) {
-				fatal_error(tls, __ccgo_ts+20154, libc.VaList(bp+16, int64(int64(nPCache))*int64(int64(szPCache))))
+				fatal_error(tls, __ccgo_ts+20154, libc.VaList(bp+16, int64(nPCache)*int64(szPCache)))
 			}
 		}
 		rc = libsqlite3.Xsqlite3_config(tls, int32(SQLITE_CONFIG_PAGECACHE), libc.VaList(bp+16, pPCache, szPCache, nPCache))
